==> 3 Model training.ipynb <==
> In this notebook a number of models are trained on the UCI data. Required libraries and the UCI dataset are first imported.

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import pickle
from sklearn.model_selection import train_test_split, GridSearchCV, StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.svm import LinearSVC, SVC
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import roc_curve, auc, confusion_matrix, f1_score, classification_report
from sklearn.metrics import make_scorer, precision_score, recall_score
from scipy.special import expit, logit

df = pd.read_csv('EEG_UCI_dataset_powers.csv')
df

df = df.rename(columns={'Unnamed: 0': 'subject'})
df['status'] = (df['subject'].str.slice(start=3, stop=4) == "a").astype(int)
df.tail()

print("Alcoholic subjects:", len(df.status.loc[df.status ==1]))
print("Control subjects:", len(df.status.loc[df.status == 0]))
print("Proportion of alcoholic subjects:", round(len(df.status.loc[df.status == 1]) / (len(df.status.loc[df.status == 1]) + len(df.status.loc[df.status == 0])), 3))

> The dataset is then split into training and test sets and pipelines are prepared.

y = df.status
X = df.drop(['subject', 'status'], axis=1)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=1234, stratify=df.status)

print(len(X_train), len(X_test), len(y_train), len(y_test))

pipelines = {
    'l1': make_pipeline(StandardScaler(), LogisticRegression(penalty='l1', solver='liblinear', random_state=123)),
    'l2': make_pipeline(StandardScaler(), LogisticRegression(penalty='l2', random_state=123)),
    'rf': make_pipeline(StandardScaler(), RandomForestClassifier(random_state=123)),
    'gb': make_pipeline(StandardScaler(), GradientBoostingClassifier(random_state=123)),
    'svm_linear': make_pipeline(StandardScaler(), LinearSVC(random_state=123)),
    'svm_rbf': make_pipeline(StandardScaler(), SVC(kernel='rbf', random_state=123))
}

l1_hyperparameters = {'logisticregression__C': np.logspace(-3, 3, 13)}
l2_hyperparameters = {'logisticregression__C': np.logspace(-3, 3, 13)}
rf_hyperparameters = {
    'randomforestclassifier__n_estimators': [25, 50, 75, 100, 150, 200],
    'randomforestclassifier__max_features': [None, 'sqrt',0.2, 0.33]
}
gb_hyperparameters = {
    'gradientboostingclassifier__n_estimators': [25, 50, 75, 100, 150, 200],
    'gradientboostingclassifier__learning_rate': [0.05, 0.1, 0.2],
    'gradientboostingclassifier__max_depth': [1, 3, 5]
}
svm_linear_hyperparameters = {'linearsvc__C': [0.1, 1, 10, 100]}
svm_rbf_hyperparameters = {
    'svc__C': [0.1, 1, 10, 100],
    'svc__gamma': [0.0001, 0.001, 0.01, 0.1, 1, 10]
}

hyperparameters = {
    'l1': l1_hyperparameters,
    'l2': l2_hyperparameters,
    'rf': rf_hyperparameters,
    'gb': gb_hyperparameters,
    'svm_linear': svm_linear_hyperparameters,
    'svm_rbf': svm_rbf_hyperparameters
}

> The models are then trained and fitted using 3-fold cross validation to find optimal hyperparameters. Accuracy scores are shown, obtained from the best cross-validated models using the training data.

fitted_models = {}
for i in ['l1', 'l2', 'rf', 'gb', 'svm_linear', 'svm_rbf']:
    model = GridSearchCV(pipelines[i], hyperparameters[i], cv=3, scoring='f1', n_jobs=-1)
    model.fit(X_train, y_train)
    fitted_models[i] = model
    print(i, 'has been fitted')

for name, model in fitted_models.items():
    print(name, model.best_score_)

> The models are then used to make predictions on the test dataset, with metrics displayed in detail below. The confusion matrix in each case is formatted as follows:  
> 
> &nbsp; &nbsp; \[[ true negatives | false positives ]  
> &nbsp; &nbsp; &nbsp;[ false negatives | true positives ]]  
> 
> **L2-regularised logistic regression** is the winning model for both accuracy and (class 1) F1 score.

print("Scores for each model:\n")
for name, model in fitted_models.items():
    if name in ['l1', 'l2', 'rf', 'gb']:
        pred = model.predict_proba(X_test)
        pred = pred[:, 1]
        y_pred = [int(p > 0.5) for p in pred]
    else:
        pred = model.decision_function(X_test)
        y_pred = [int(p > 0) for p in pred]
    fpr, tpr, thresholds = roc_curve(y_test, pred)
    print(name, "- AUROC:", auc(fpr, tpr), " F1 score:", f1_score(y_test, y_pred))
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred))
    print()

> The next cells use **nested cross validation**. This was tried because of the relatively small size of the dataset. In this process, there are 4 outer loops of validation. Within each outer loop, an inner loop of 3-fold cross validation is performed as before, and the best model for each algorithm is tested on a hold-out set. Because nested cross validation does not produce a single best fitted model, the aim is to find the best *process* from amongst the various algorithms tested, based on average metrics across the 4 outer loops.
> 
> Additionally here, **threshold optimisation** is applied. In an attempt to improve upon the default threshold of 0.5 for the prediction probability (or 0 for decision function, which becomes 0.5 when the logistic function is applied), thresholds from 0.01 to 0.99 are tested on the models to find the best average F1 score across the 3 inner loops of cross validation. Note that a different best threshold might be produced for each outer loop of validation, so the average of these 4 results, obtained by testing the best model on the hold-out set in each case, is shown in the summary at the end of this cell's output.

summary = "Mean scores of best models:\n\n"
for i in ['l1', 'l2', 'rf', 'gb', 'svm_linear', 'svm_rbf']:
    print("****", i.upper(), "****\n")
    cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
    outer_results_def_thr = list()
    outer_results_best_thr = list()
    outloop = 0
    for train_ix, test_ix in cv_outer.split(X, y):
        outloop += 1
        X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
        y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
        cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
        clf = GridSearchCV(pipelines[i], hyperparameters[i], cv=cv_inner, scoring='f1', n_jobs=-1)
        clf.fit(X_train, y_train)
        f1_array = np.zeros(99)
        for itrain_ix, itest_ix in cv_inner.split(X_train, y_train):
            iX_train, iX_test = X_train.iloc[itrain_ix, :], X_train.iloc[itest_ix, :]
            iy_train, iy_test = y_train.iloc[itrain_ix], y_train.iloc[itest_ix]
            model = pipelines[i]
            model.set_params(**clf.best_params_)
            model.fit(iX_train, iy_train)
            if i in ['l1', 'l2', 'rf', 'gb']:
                pred = model.predict_proba(iX_test)
                pred = pred[:, 1]
            else:
                pred = model.decision_function(iX_test)
            for thr in range(1, 100):
                if i in ['l1', 'l2', 'rf', 'gb']:
                    y_pred = [int(p > thr / 100) for p in pred]
                else:
                    y_pred = [int(expit(p) > thr / 100) for p in pred]
                f1_array[int(thr - 1)] += f1_score(iy_test, y_pred, zero_division=0)
        best_threshold = (np.argmax(f1_array) + 1) / 100
        if i in ['svm_linear', 'svm_rbf']:
            best_threshold = logit(best_threshold)
        if i in ['l1', 'l2', 'rf', 'gb']:
            pred = clf.predict_proba(X_test)
            pred = pred[:, 1]
            y_pred = [int(p > 0.5) for p in pred]
        else:
            pred = clf.decision_function(X_test)
            y_pred = [int(p > 0) for p in pred]
        print("Outer loop " + str(outloop) + ":\n")
        print("Default threshold")
        fpr, tpr, thresholds = roc_curve(y_test, pred)
        print(" F1 score:", f1_score(y_test, y_pred))
        print(confusion_matrix(y_test, y_pred))
        print()
        outer_results_def_thr.append({'f1': f1_score(y_test, y_pred, zero_division=0), 'AUROC': auc(fpr, tpr),
                                      'sensitivity': recall_score(y_test, y_pred), 'specificity': recall_score(y_test, y_pred, pos_label=0),
                                      'PPV': precision_score(y_test, y_pred, zero_division=0), 'NPV': precision_score(y_test, y_pred, pos_label=0, zero_division=0)})
        print("Best threshold:", best_threshold)
        if i in ['l1', 'l2', 'rf', 'gb']:
            y_pred = [int(p > best_threshold) for p in pred]
        else:
            y_pred = [int(expit(p) > best_threshold) for p in pred]
        fpr, tpr, thresholds = roc_curve(y_test, pred)
        print(" F1 score:", f1_score(y_test, y_pred))
        print(confusion_matrix(y_test, y_pred))
        print()
        outer_results_best_thr.append({'f1': f1_score(y_test, y_pred, zero_division=0), 'AUROC': auc(fpr, tpr),
                                      'sensitivity': recall_score(y_test, y_pred), 'specificity': recall_score(y_test, y_pred, pos_label=0),
                                      'PPV': precision_score(y_test, y_pred, zero_division=0), 'NPV': precision_score(y_test, y_pred, pos_label=0, zero_division=0)})
    summary = summary + i + "\n  Default threshold\n    AUROC: " + str(np.array([dict['AUROC'] for dict in outer_results_def_thr]).mean()) + "  F1 score: " + str(np.array([dict['f1'] for dict in outer_results_def_thr]).mean())
    summary = summary + "\n    Sensitivity: " + str(np.array([dict['sensitivity'] for dict in outer_results_def_thr]).mean())[:6] + "  Specificity: " + str(np.array([dict['specificity'] for dict in outer_results_def_thr]).mean())[:6]
    summary = summary + "\n    PPV: " + str(np.array([dict['PPV'] for dict in outer_results_def_thr]).mean())[:6] + "  NPV: " + str(np.array([dict['NPV'] for dict in outer_results_def_thr]).mean())[:6]
    summary = summary + "\n  Best thresholds\n    AUROC: " + str(np.array([dict['AUROC'] for dict in outer_results_best_thr]).mean()) + "  F1 score: " + str(np.array([dict['f1'] for dict in outer_results_best_thr]).mean())
    summary = summary + "\n    Sensitivity: " + str(np.array([dict['sensitivity'] for dict in outer_results_best_thr]).mean())[:6] + "  Specificity: " + str(np.array([dict['specificity'] for dict in outer_results_best_thr]).mean())[:6]
    summary = summary + "\n    PPV: " + str(np.array([dict['PPV'] for dict in outer_results_best_thr]).mean())[:6] + "  NPV: " + str(np.array([dict['NPV'] for dict in outer_results_best_thr]).mean())[:6] + "\n"
print(summary) 

> Nested cross validation is used again, this time using area under receiver operating curve (AUROC) as the optimising variable for hyperparameters, and geometric mean to optimise thresholds.

def g_mean(y_test, y_pred):
    g_mean = np.sqrt(recall_score(y_test, y_pred) * recall_score(y_test, y_pred, pos_label=0))
    return g_mean

summary = "Mean scores of best models:\n\n"

for i in ['l1', 'l2', 'rf', 'gb', 'svm_linear', 'svm_rbf']:
    print("****", i.upper(), "****\n")
    cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
    outer_results_def_thr = list()
    outer_results_best_thr = list()
    outloop = 0
    for train_ix, test_ix in cv_outer.split(X, y):
        outloop += 1
        X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
        y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
        cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
        clf = GridSearchCV(pipelines[i], hyperparameters[i], cv=cv_inner, scoring='roc_auc', n_jobs=-1)
        clf.fit(X_train, y_train)
        gm_array = np.zeros(99)
        for itrain_ix, itest_ix in cv_inner.split(X_train, y_train):
            iX_train, iX_test = X_train.iloc[itrain_ix, :], X_train.iloc[itest_ix, :]
            iy_train, iy_test = y_train.iloc[itrain_ix], y_train.iloc[itest_ix]
            model = pipelines[i]
            model.set_params(**clf.best_params_)
            model.fit(iX_train, iy_train)
            if i in ['l1', 'l2', 'rf', 'gb']:
                pred = model.predict_proba(iX_test)
                pred = pred[:, 1]
            else:
                pred = model.decision_function(iX_test)
            for thr in range(1, 100):
                if i in ['l1', 'l2', 'rf', 'gb']:
                    y_pred = [int(p > thr / 100) for p in pred]
                else:
                    y_pred = [int(expit(p) > thr / 100) for p in pred]
                gm_array[int(thr - 1)] += g_mean(iy_test, y_pred)
        best_threshold = (np.argmax(gm_array) + 1) / 100
        if i in ['svm_linear', 'svm_rbf']:
            best_threshold = logit(best_threshold)
        if i in ['l1', 'l2', 'rf', 'gb']:
            pred = clf.predict_proba(X_test)
            pred = pred[:, 1]
            y_pred = [int(p > 0.5) for p in pred]
        else:
            pred = clf.decision_function(X_test)
            y_pred = [int(p > 0) for p in pred]
        print("Outer loop " + str(outloop) + ":\n")
        print("Default threshold")
        fpr, tpr, thresholds = roc_curve(y_test, pred)
        print(" G-mean:", g_mean(y_test, y_pred))
        print(confusion_matrix(y_test, y_pred))
        print()
        outer_results_def_thr.append({'f1': f1_score(y_test, y_pred, zero_division=0), 'AUROC': auc(fpr, tpr), 'g-mean': g_mean(y_test, y_pred),
                                      'sensitivity': recall_score(y_test, y_pred), 'specificity': recall_score(y_test, y_pred, pos_label=0),
                                      'PPV': precision_score(y_test, y_pred, zero_division=0), 'NPV': precision_score(y_test, y_pred, pos_label=0, zero_division=0)})
        print("Best threshold:", best_threshold)
        if i in ['l1', 'l2', 'rf', 'gb']:
            y_pred = [int(p > best_threshold) for p in pred]
        else:
            y_pred = [int(expit(p) > best_threshold) for p in pred]
        fpr, tpr, thresholds = roc_curve(y_test, pred)
        print(" G-mean:", g_mean(y_test, y_pred))
        print(confusion_matrix(y_test, y_pred))
        print()
        outer_results_best_thr.append({'f1': f1_score(y_test, y_pred, zero_division=0), 'AUROC': auc(fpr, tpr), 'g-mean': g_mean(y_test, y_pred),
                                      'sensitivity': recall_score(y_test, y_pred), 'specificity': recall_score(y_test, y_pred, pos_label=0),
                                      'PPV': precision_score(y_test, y_pred, zero_division=0), 'NPV': precision_score(y_test, y_pred, pos_label=0, zero_division=0)})
    summary = summary + i + "\n  Default threshold\n    AUROC: " + str(np.array([dict['AUROC'] for dict in outer_results_def_thr]).mean()) + "  F1 score: " + str(np.array([dict['f1'] for dict in outer_results_def_thr]).mean())
    summary = summary + "\n    Sensitivity: " + str(np.array([dict['sensitivity'] for dict in outer_results_def_thr]).mean())[:6] + "  Specificity: " + str(np.array([dict['specificity'] for dict in outer_results_def_thr]).mean())[:6]
    summary = summary + "\n    PPV: " + str(np.array([dict['PPV'] for dict in outer_results_def_thr]).mean())[:6] + "  NPV: " + str(np.array([dict['NPV'] for dict in outer_results_def_thr]).mean())[:6]
    summary = summary + "\n    G-mean: " + str(np.array([dict['g-mean'] for dict in outer_results_def_thr]).mean())
    summary = summary + "\n  Best thresholds\n    AUROC: " + str(np.array([dict['AUROC'] for dict in outer_results_best_thr]).mean()) + "  F1 score: " + str(np.array([dict['f1'] for dict in outer_results_best_thr]).mean())
    summary = summary + "\n    Sensitivity: " + str(np.array([dict['sensitivity'] for dict in outer_results_best_thr]).mean())[:6] + "  Specificity: " + str(np.array([dict['specificity'] for dict in outer_results_best_thr]).mean())[:6]
    summary = summary + "\n    PPV: " + str(np.array([dict['PPV'] for dict in outer_results_best_thr]).mean())[:6] + "  NPV: " + str(np.array([dict['NPV'] for dict in outer_results_best_thr]).mean())[:6]
    summary = summary + "\n    G-mean: " + str(np.array([dict['g-mean'] for dict in outer_results_best_thr]).mean()) + "\n"

print(summary) 

> The most favourable result overall appears to be the one using **L2-regularised logistic regression** with F1 score as the optimising variable. This process is then applied anew using 3-fold cross validation on the full dataset in order to produce a fitted model and a best threshold specific to this model.

i = 'l2'
results_best_thr = list()
cv = StratifiedKFold(n_splits=3, shuffle=True, random_state=2)
clf = GridSearchCV(pipelines[i], hyperparameters[i], cv=cv, scoring='f1', n_jobs=-1)
clf.fit(X, y)
f1_array = np.zeros(99)
for train_ix, test_is in cv.split(X, y):
    X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
    y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
    model = pipelines[i]
    model.set_params(**clf.best_params_)
    model.fit(X_train, y_train)
    if i in ['l1', 'l2', 'rf', 'gb']:
        pred = model.predict_proba(X_test)
        pred = pred[:, 1]
    else:
        pred = model.decision_function(X_test)
    for thr in range(1, 100):
        if i in ['l1', 'l2', 'rf', 'gb']:
            y_pred = [int(p > thr / 100) for p in pred]
        else:
            y_pred = [int(expit(p) > thr / 100) for p in pred]
        f1_array[int(thr - 1)] += f1_score(y_test, y_pred)
best_threshold = (np.argmax(f1_array) + 1) / 100
if i in ['svm_linear', 'svm_rbf']:
    best_threshold = logit(best_threshold)
fitted_model_ncv = clf
best_thr_ncv = best_threshold
print("Best threshold for", i.upper(), "model =", best_thr_ncv)

> **WITHOUT ABSOLUTE POWERS**
> 
> In the following sections I've repeated all the above processes with absolute powers removed from the dataset to leave only  relative powers and lateralisation variables. The presumption I've made is that different testing conditions such as different machinery and different ground lead placement may lead to variation in absolute powers, whereas if these are all varied in a similar manner then relative powers should be more comparable across testing locations.

np.array(df.columns)

y = df.status
X = df.drop(['subject', 'Fp1a delta', 'Fp1a theta', 'Fp1a alpha', 'Fp1a beta',
       'Fp1a gamma', 'Fp2a delta', 'Fp2a theta', 'Fp2a alpha',
       'Fp2a beta', 'Fp2a gamma', 'F3a delta', 'F3a theta', 'F3a alpha',
       'F3a beta', 'F3a gamma', 'F4a delta', 'F4a theta', 'F4a alpha',
       'F4a beta', 'F4a gamma', 'F7a delta', 'F7a theta', 'F7a alpha',
       'F7a beta', 'F7a gamma', 'F8a delta', 'F8a theta', 'F8a alpha',
       'F8a beta', 'F8a gamma', 'C3a delta', 'C3a theta', 'C3a alpha',
       'C3a beta', 'C3a gamma', 'C4a delta', 'C4a theta', 'C4a alpha',
       'C4a beta', 'C4a gamma', 'P3a delta', 'P3a theta', 'P3a alpha',
       'P3a beta', 'P3a gamma', 'P4a delta', 'P4a theta', 'P4a alpha',
       'P4a beta', 'P4a gamma', 'O1a delta', 'O1a theta', 'O1a alpha',
       'O1a beta', 'O1a gamma', 'O2a delta', 'O2a theta', 'O2a alpha',
       'O2a beta', 'O2a gamma', 'status'], axis=1)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=1234, stratify=df.status)

> Cross validation is performed as before to produce a fitted model for each algorithm.

fitted_models_no_ap = {}
for i in ['l1', 'l2', 'rf', 'gb', 'svm_linear', 'svm_rbf']:
    model = GridSearchCV(pipelines[i], hyperparameters[i], cv=3, scoring='f1', n_jobs=-1)
    model.fit(X_train, y_train)
    fitted_models_no_ap[i] = model
    print(i, 'has been fitted')

for name, model in fitted_models_no_ap.items():
    print(name, model.best_score_)

> Choosing the best model below is somewhat subjective. The support vector machine using radial basis function kernel (svm_rbf) gives the best F1 score but the model is useless because every subject is predicted as being alcoholic. I've chosen the **support vector machine** using a **linear kernel (svm_linear)** as this appears to be the most clinically useful model in terms of correctly identifying a non-alcoholic subject out of the 4 models with an equal highest accuracy of 0.68.

print("Scores for each model:\n")
for name, model in fitted_models_no_ap.items():
    if name in ['l1', 'l2', 'rf', 'gb']:
        pred = model.predict_proba(X_test)
        pred = [p[1] for p in pred]
        y_pred = [int(p > 0.5) for p in pred]
    else:
        pred = model.decision_function(X_test)
        y_pred = [int(p > 0) for p in pred]
    fpr, tpr, thresholds = roc_curve(y_test, pred)
    print(name, "- AUROC:", auc(fpr, tpr), " F1 score:", f1_score(y_test, y_pred))
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred))
    print()

> Nested cross validation is performed as previously.

summary = "Mean scores of best models:\n\n"
for i in ['l1', 'l2', 'rf', 'gb', 'svm_linear', 'svm_rbf']:
    print("****", i.upper(), "****\n")
    cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
    outer_results_def_thr = list()
    outer_results_best_thr = list()
    outloop = 0
    for train_ix, test_ix in cv_outer.split(X, y):
        outloop += 1
        X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
        y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
        cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
        clf = GridSearchCV(pipelines[i], hyperparameters[i], cv=cv_inner, scoring='f1', n_jobs=-1)
        clf.fit(X_train, y_train)
        f1_array = np.zeros(99)
        for itrain_ix, itest_ix in cv_inner.split(X_train, y_train):
            iX_train, iX_test = X_train.iloc[itrain_ix, :], X_train.iloc[itest_ix, :]
            iy_train, iy_test = y_train.iloc[itrain_ix], y_train.iloc[itest_ix]
            model = pipelines[i]
            model.set_params(**clf.best_params_)
            model.fit(iX_train, iy_train)
            if i in ['l1', 'l2', 'rf', 'gb']:
                pred = model.predict_proba(iX_test)
                pred = pred[:, 1]
            else:
                pred = model.decision_function(iX_test)
            for thr in range(1, 100):
                if i in ['l1', 'l2', 'rf', 'gb']:
                    y_pred = [int(p > thr / 100) for p in pred]
                else:
                    y_pred = [int(expit(p) > thr / 100) for p in pred]
                f1_array[int(thr - 1)] += f1_score(iy_test, y_pred, zero_division=0)
        best_threshold = (np.argmax(f1_array) + 1) / 100
        if i in ['svm_linear', 'svm_rbf']:
            best_threshold = logit(best_threshold)
        if i in ['l1', 'l2', 'rf', 'gb']:
            pred = clf.predict_proba(X_test)
            pred = pred[:, 1]
            y_pred = [int(p > 0.5) for p in pred]
        else:
            pred = clf.decision_function(X_test)
            y_pred = [int(p > 0) for p in pred]
        print("Outer loop " + str(outloop) + ":\n")
        print("Default threshold")
        fpr, tpr, thresholds = roc_curve(y_test, pred)
        print(" F1 score:", f1_score(y_test, y_pred))
        print(confusion_matrix(y_test, y_pred))
        print()
        outer_results_def_thr.append({'f1': f1_score(y_test, y_pred, zero_division=0), 'AUROC': auc(fpr, tpr),
                                      'sensitivity': recall_score(y_test, y_pred), 'specificity': recall_score(y_test, y_pred, pos_label=0),
                                      'PPV': precision_score(y_test, y_pred, zero_division=0), 'NPV': precision_score(y_test, y_pred, pos_label=0, zero_division=0)})
        print("Best threshold:", best_threshold)
        if i in ['l1', 'l2', 'rf', 'gb']:
            y_pred = [int(p > best_threshold) for p in pred]
        else:
            y_pred = [int(expit(p) > best_threshold) for p in pred]
        fpr, tpr, thresholds = roc_curve(y_test, pred)
        print(" F1 score:", f1_score(y_test, y_pred))
        print(confusion_matrix(y_test, y_pred))
        print()
        outer_results_best_thr.append({'f1': f1_score(y_test, y_pred, zero_division=0), 'AUROC': auc(fpr, tpr),
                                      'sensitivity': recall_score(y_test, y_pred), 'specificity': recall_score(y_test, y_pred, pos_label=0),
                                      'PPV': precision_score(y_test, y_pred, zero_division=0), 'NPV': precision_score(y_test, y_pred, pos_label=0, zero_division=0)})
    summary = summary + i + "\n  Default threshold\n    AUROC: " + str(np.array([dict['AUROC'] for dict in outer_results_def_thr]).mean()) + "  F1 score: " + str(np.array([dict['f1'] for dict in outer_results_def_thr]).mean())
    summary = summary + "\n    Sensitivity: " + str(np.array([dict['sensitivity'] for dict in outer_results_def_thr]).mean())[:6] + "  Specificity: " + str(np.array([dict['specificity'] for dict in outer_results_def_thr]).mean())[:6]
    summary = summary + "\n    PPV: " + str(np.array([dict['PPV'] for dict in outer_results_def_thr]).mean())[:6] + "  NPV: " + str(np.array([dict['NPV'] for dict in outer_results_def_thr]).mean())[:6]
    summary = summary + "\n  Best thresholds\n    AUROC: " + str(np.array([dict['AUROC'] for dict in outer_results_best_thr]).mean()) + "  F1 score: " + str(np.array([dict['f1'] for dict in outer_results_best_thr]).mean())
    summary = summary + "\n    Sensitivity: " + str(np.array([dict['sensitivity'] for dict in outer_results_best_thr]).mean())[:6] + "  Specificity: " + str(np.array([dict['specificity'] for dict in outer_results_best_thr]).mean())[:6]
    summary = summary + "\n    PPV: " + str(np.array([dict['PPV'] for dict in outer_results_best_thr]).mean())[:6] + "  NPV: " + str(np.array([dict['NPV'] for dict in outer_results_best_thr]).mean())[:6] + "\n"
print(summary) 

def g_mean(y_test, y_pred):
    g_mean = np.sqrt(recall_score(y_test, y_pred) * recall_score(y_test, y_pred, pos_label=0))
    return g_mean

summary = "Mean scores of best models:\n\n"

for i in ['l1', 'l2', 'rf', 'gb', 'svm_linear', 'svm_rbf']:
    print("****", i.upper(), "****\n")
    cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
    outer_results_def_thr = list()
    outer_results_best_thr = list()
    outloop = 0
    for train_ix, test_ix in cv_outer.split(X, y):
        outloop += 1
        X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
        y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
        cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
        clf = GridSearchCV(pipelines[i], hyperparameters[i], cv=cv_inner, scoring='roc_auc', n_jobs=-1)
        clf.fit(X_train, y_train)
        gm_array = np.zeros(99)
        for itrain_ix, itest_ix in cv_inner.split(X_train, y_train):
            iX_train, iX_test = X_train.iloc[itrain_ix, :], X_train.iloc[itest_ix, :]
            iy_train, iy_test = y_train.iloc[itrain_ix], y_train.iloc[itest_ix]
            model = pipelines[i]
            model.set_params(**clf.best_params_)
            model.fit(iX_train, iy_train)
            if i in ['l1', 'l2', 'rf', 'gb']:
                pred = model.predict_proba(iX_test)
                pred = pred[:, 1]
            else:
                pred = model.decision_function(iX_test)
            for thr in range(1, 100):
                if i in ['l1', 'l2', 'rf', 'gb']:
                    y_pred = [int(p > thr / 100) for p in pred]
                else:
                    y_pred = [int(expit(p) > thr / 100) for p in pred]
                gm_array[int(thr - 1)] += g_mean(iy_test, y_pred)
        best_threshold = (np.argmax(gm_array) + 1) / 100
        if i in ['svm_linear', 'svm_rbf']:
            best_threshold = logit(best_threshold)
        if i in ['l1', 'l2', 'rf', 'gb']:
            pred = clf.predict_proba(X_test)
            pred = pred[:, 1]
            y_pred = [int(p > 0.5) for p in pred]
        else:
            pred = clf.decision_function(X_test)
            y_pred = [int(p > 0) for p in pred]
        print("Outer loop " + str(outloop) + ":\n")
        print("Default threshold")
        fpr, tpr, thresholds = roc_curve(y_test, pred)
        print(" G-mean:", g_mean(y_test, y_pred))
        print(confusion_matrix(y_test, y_pred))
        print()
        outer_results_def_thr.append({'f1': f1_score(y_test, y_pred, zero_division=0), 'AUROC': auc(fpr, tpr), 'g-mean': g_mean(y_test, y_pred),
                                      'sensitivity': recall_score(y_test, y_pred), 'specificity': recall_score(y_test, y_pred, pos_label=0),
                                      'PPV': precision_score(y_test, y_pred, zero_division=0), 'NPV': precision_score(y_test, y_pred, pos_label=0, zero_division=0)})
        print("Best threshold:", best_threshold)
        if i in ['l1', 'l2', 'rf', 'gb']:
            y_pred = [int(p > best_threshold) for p in pred]
        else:
            y_pred = [int(expit(p) > best_threshold) for p in pred]
        fpr, tpr, thresholds = roc_curve(y_test, pred)
        print(" G-mean:", g_mean(y_test, y_pred))
        print(confusion_matrix(y_test, y_pred))
        print()
        outer_results_best_thr.append({'f1': f1_score(y_test, y_pred, zero_division=0), 'AUROC': auc(fpr, tpr), 'g-mean': g_mean(y_test, y_pred),
                                      'sensitivity': recall_score(y_test, y_pred), 'specificity': recall_score(y_test, y_pred, pos_label=0),
                                      'PPV': precision_score(y_test, y_pred, zero_division=0), 'NPV': precision_score(y_test, y_pred, pos_label=0, zero_division=0)})
    summary = summary + i + "\n  Default threshold\n    AUROC: " + str(np.array([dict['AUROC'] for dict in outer_results_def_thr]).mean()) + "  F1 score: " + str(np.array([dict['f1'] for dict in outer_results_def_thr]).mean())
    summary = summary + "\n    Sensitivity: " + str(np.array([dict['sensitivity'] for dict in outer_results_def_thr]).mean())[:6] + "  Specificity: " + str(np.array([dict['specificity'] for dict in outer_results_def_thr]).mean())[:6]
    summary = summary + "\n    PPV: " + str(np.array([dict['PPV'] for dict in outer_results_def_thr]).mean())[:6] + "  NPV: " + str(np.array([dict['NPV'] for dict in outer_results_def_thr]).mean())[:6]
    summary = summary + "\n    G-mean: " + str(np.array([dict['g-mean'] for dict in outer_results_def_thr]).mean())
    summary = summary + "\n  Best thresholds\n    AUROC: " + str(np.array([dict['AUROC'] for dict in outer_results_best_thr]).mean()) + "  F1 score: " + str(np.array([dict['f1'] for dict in outer_results_best_thr]).mean())
    summary = summary + "\n    Sensitivity: " + str(np.array([dict['sensitivity'] for dict in outer_results_best_thr]).mean())[:6] + "  Specificity: " + str(np.array([dict['specificity'] for dict in outer_results_best_thr]).mean())[:6]
    summary = summary + "\n    PPV: " + str(np.array([dict['PPV'] for dict in outer_results_best_thr]).mean())[:6] + "  NPV: " + str(np.array([dict['NPV'] for dict in outer_results_best_thr]).mean())[:6]
    summary = summary + "\n    G-mean: " + str(np.array([dict['g-mean'] for dict in outer_results_best_thr]).mean()) + "\n"

print(summary) 

> Again the choice of a best process is subjective, and **L1-regularised linear regression** optimised for AUROC (and with thresholds optimised for geometric mean) is chosen on the basis of having a sensitivity and a specificity which are both of some clinical use.

i = 'l1'
results_best_thr = list()
cv = StratifiedKFold(n_splits=3, shuffle=True, random_state=2)
clf = GridSearchCV(pipelines[i], hyperparameters[i], cv=cv, scoring='roc_auc', n_jobs=-1)
clf.fit(X, y)
gm_array = np.zeros(99)
for train_ix, test_is in cv.split(X, y):
    X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
    y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
    model = pipelines[i]
    model.set_params(**clf.best_params_)
    model.fit(X_train, y_train)
    if i in ['l1', 'l2', 'rf', 'gb']:
        pred = model.predict_proba(X_test)
        pred = pred[:, 1]
    else:
        pred = model.decision_function(X_test)
    for thr in range(1, 100):
        if i in ['l1', 'l2', 'rf', 'gb']:
            y_pred = [int(p > thr / 100) for p in pred]
        else:
            y_pred = [int(expit(p) > thr / 100) for p in pred]
        gm_array[int(thr - 1)] += g_mean(y_test, y_pred)
best_threshold = (np.argmax(f1_array) + 1) / 100
if i in ['svm_linear', 'svm_rbf']:
    best_threshold = logit(best_threshold)
fitted_model_no_ap_ncv = clf
best_thr_no_ap_ncv = best_threshold
print("Best threshold for", i.upper(), "model =", best_thr_no_ap_ncv)

> Hyperparameters of the best models are shown below, then the models and best thresholds are saved for further use using the Pickle module.

fitted_models['l2'].best_estimator_

fitted_model_ncv.best_estimator_

fitted_models_no_ap['svm_linear'].best_estimator_

fitted_model_no_ap_ncv.best_estimator_

with open("eeg_model.pkl", "wb") as f:
    pickle.dump(fitted_models['l2'].best_estimator_, f)
with open("eeg_model_ncv.pkl", "wb") as f:
    pickle.dump(fitted_model_ncv.best_estimator_, f)
with open("eeg_best_thr_ncv.pkl", "wb") as f:
    pickle.dump(best_thr_ncv, f)
with open("eeg_model_no_ap.pkl", "wb") as f:
    pickle.dump(fitted_models_no_ap['svm_linear'].best_estimator_, f)
with open("eeg_model_no_ap_ncv.pkl", "wb") as f:
    pickle.dump(fitted_model_no_ap_ncv.best_estimator_, f)
with open("eeg_best_thr_no_ap_ncv.pkl", "wb") as f:
    pickle.dump(best_thr_no_ap_ncv, f)
