==> 4 Deep model training.ipynb <==
> This notebook goes through a similar process as the previous one to train models on the UCI dataset, this time with deep learning models. Required libraries and the dataset are first imported.

import numpy as np
import pandas as pd
import pickle
import joblib
from sklearn.model_selection import train_test_split, GridSearchCV, RandomizedSearchCV, StratifiedKFold
from sklearn.metrics import roc_curve, auc, confusion_matrix, f1_score, classification_report
from sklearn.metrics import make_scorer, precision_score, recall_score
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler
from keras.models import Sequential
from keras.layers import Dense, Dropout
from keras.constraints import max_norm
from keras.wrappers.scikit_learn import KerasClassifier

df = pd.read_csv('EEG_UCI_dataset_powers.csv')
df = df.rename(columns={'Unnamed: 0': 'subject'})
df['status'] = (df['subject'].str.slice(start=3, stop=4) == "a").astype(int)
df.tail()

print("Alcoholic subjects:", len(df.status.loc[df.status ==1]))
print("Control subjects:", len(df.status.loc[df.status == 0]))
print("Proportion of alcoholic subjects:", round(len(df.status.loc[df.status == 1]) / (len(df.status.loc[df.status == 1]) + len(df.status.loc[df.status == 0])), 3))

> The dataset is split into training and test sets, then the KerasClassifier module is used in order to set up a pipeline for a grid search across hyperparameters using cross validation.

y = df.status
X = df.drop(['subject', 'status'], axis=1)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=1234, stratify=df.status)
print(len(X_train), len(X_test), len(y_train), len(y_test))

def create_model(neurons1=100, neurons2=30, dropout1=0, dropout2=0, constraint=0):
    model = Sequential()
    model.add(Dense(neurons1, input_dim=150, activation='relu', kernel_constraint=max_norm(constraint)))
    model.add(Dropout(dropout1))
    model.add(Dense(neurons2, activation='relu', kernel_constraint=max_norm(constraint)))
    model.add(Dropout(dropout2))
    model.add(Dense(1, activation='sigmoid'))
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
    return model

pipeline = Pipeline([('scaler', StandardScaler()), ('nn', KerasClassifier(build_fn=create_model, verbose=0))])

> A number of models are then created using GridSearchCV or RandomizedSearchCV and tested on the test set. The choice of hyperparameters is then refined each time and the most favourable model is saved. (Note that this process is quite time-consuming and was therefore run over a number of sessions, as is apparent from the output numbers.)

# MODEL 1
np.random.seed(12)
batch_size = [10, 20, 40]
epochs = [10, 20, 40, 70, 100]
neurons1 = [50, 100, 150, 200, 250]
neurons2 = [10, 20, 30, 40, 50]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2)
clf = GridSearchCV(estimator=pipeline, param_grid=param_grid, cv=3, n_jobs=-1, verbose=1)
clf.fit(X_train, y_train)

y_pred = (clf.predict(X_test) > 0.5).astype(int)
print(clf.best_score_, clf.best_params_)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

# MODEL 2
np.random.seed(12)
batch_size = [10, 20, 40]
epochs = [10, 20, 40]
neurons1 = [50, 100, 150, 200, 250]
neurons2 = [30, 40, 50, 60]
dropout1 = [0, 0.1, 0.2]
dropout2 = [0, 0.1, 0.2]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2)
clf = GridSearchCV(estimator=pipeline, param_grid=param_grid, cv=3, n_jobs=6, verbose=1)
clf.fit(X_train, y_train)

y_pred = (clf.predict(X_test) > 0.5).astype(int)
print(clf.best_score_, clf.best_params_)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

# MODEL 3
np.random.seed(12)
batch_size = [10, 20, 30, 40]
epochs = [8, 10, 12, 14, 16, 18, 20]
neurons1 = [50, 75, 100, 125, 150, 175, 200]
neurons2 = [30, 40, 50, 60]
dropout1 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
dropout2 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
constraint = [0, 1, 2, 3, 4, 5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
clf = RandomizedSearchCV(estimator=pipeline, param_distributions=param_grid, cv=3, n_jobs=3, verbose=1, n_iter=1000)
clf.fit(X_train, y_train)

y_pred = (clf.predict(X_test) > 0.5).astype(int)
print(clf.best_score_, clf.best_params_)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

# MODEL 4
np.random.seed(12)
batch_size = [30, 35, 40, 45, 50]
epochs = [6, 7, 8, 9, 10, 11, 12]
neurons1 = [25, 37, 50, 75, 100]
neurons2 = [30, 40, 50, 60]
dropout1 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
dropout2 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
constraint = [0, 1, 2, 3, 4, 5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
clf = RandomizedSearchCV(estimator=pipeline, param_distributions=param_grid, cv=3, n_jobs=3, verbose=1, n_iter=1000, random_state=1)
clf.fit(X_train, y_train)

y_pred = (clf.predict(X_test) > 0.5).astype(int)
print(clf.best_score_, clf.best_params_)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

fitted_dl_model = clf.best_estimator_

# MODEL 5
np.random.seed(12)
batch_size = [40, 45, 50]
epochs = [10, 11, 12]
neurons1 = [67, 75, 83]
neurons2 = [45, 50, 55]
dropout1 = [0.4, 0.5]
dropout2 = [0.4, 0.5]
constraint = [3, 4]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
clf = GridSearchCV(estimator=pipeline, param_grid=param_grid, cv=3, n_jobs=6, verbose=1)
clf.fit(X_train, y_train)

y_pred = (clf.predict(X_test) > 0.5).astype(int)
print(clf.best_score_, clf.best_params_)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

> Model 1 and model 4 have produced the best results, so as model 4 has already been stored in a variable this will be saved for later use.
> 
> The entire model is not suitable for saving in a single file with either the Pickle or Joblib libraries, so it is split into two parts and saved as two files. The process for saving it this way and subsequently retrieving it is modified from https://prodevsblog.com/questions/125081/how-to-save-a-scikit-learn-pipline-with-keras-regressor-inside-to-disk/.

fitted_dl_model.named_steps['nn'].model.save('eeg_dl_model.h5')
fitted_dl_model.named_steps['nn'].model = None
joblib.dump(fitted_dl_model, 'eeg_dl_model.pkl')

> **Nested cross validation model**
>
> Nested cross validation is applied, as per the previous notebook.

# PROCESS 1
np.random.seed(12)
batch_size = [30, 35, 40, 45, 50]
epochs = [6, 7, 8, 9, 10, 11, 12]
neurons1 = [25, 37, 50, 75, 100]
neurons2 = [30, 40, 50, 60]
dropout1 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
dropout2 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
constraint = [0, 1, 2, 3, 4, 5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
outer_results = list()
for train_ix, test_ix in cv_outer.split(X, y):
    X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
    y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
    cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
    clf = RandomizedSearchCV(estimator=pipeline, param_distributions=param_grid, cv=3, n_jobs=3, verbose=1, n_iter=1000, random_state=1)
    clf.fit(X_train, y_train)
    y_pred = (clf.predict(X_test) > 0.5).astype(int)
    print("\n", clf.best_score_, clf.best_params_)
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred))
    outer_results.append({'f1': f1_score(y_test, y_pred), 'sensitivity': recall_score(y_test, y_pred),
                          'specificity': recall_score(y_test, y_pred, pos_label=0), 'PPV': precision_score(y_test, y_pred),
                          'NPV': precision_score(y_test, y_pred, pos_label=0)})
print("\nMean scores:")
for score in ['f1', 'sensitivity', 'specificity', 'PPV', 'NPV']:
    print(" ", score, ":", np.array([dict[score] for dict in outer_results]).mean())

# PROCESS 2
np.random.seed(12)
batch_size = [30, 35, 40, 45]
epochs = [6, 7, 8, 9]
neurons1 = [25, 30, 35, 40, 45, 50]
neurons2 = [30, 40, 50, 60]
dropout1 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
dropout2 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
constraint = [0, 1, 2, 3, 4, 5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
outer_results = list()
for train_ix, test_ix in cv_outer.split(X, y):
    X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
    y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
    cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
    clf = RandomizedSearchCV(estimator=pipeline, param_distributions=param_grid, cv=3, n_jobs=3, verbose=1, n_iter=1000, random_state=1)
    clf.fit(X_train, y_train)
    y_pred = clf.predict(X_test)
    print("\n", clf.best_score_, clf.best_params_)
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred))
    outer_results.append({'f1': f1_score(y_test, y_pred), 'sensitivity': recall_score(y_test, y_pred),
                          'specificity': recall_score(y_test, y_pred, pos_label=0), 'PPV': precision_score(y_test, y_pred),
                          'NPV': precision_score(y_test, y_pred, pos_label=0)})
print("\nMean scores:")
for score in ['f1', 'sensitivity', 'specificity', 'PPV', 'NPV']:
    print(" ", score, ":", np.array([dict[score] for dict in outer_results]).mean())

# PROCESS 3
np.random.seed(12)
batch_size = [40, 45]
epochs = [6, 7, 8, 9]
neurons1 = [35, 40]
neurons2 = [30, 40, 50]
dropout1 = [0.1, 0.2, 0.3, 0.4]
dropout2 = [0, 0.3, 0.5]
constraint = [2, 3, 4]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
outer_results = list()
for train_ix, test_ix in cv_outer.split(X, y):
    X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
    y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
    cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
    clf = GridSearchCV(estimator=pipeline, param_grid=param_grid, cv=3, n_jobs=6, verbose=1)
    clf.fit(X_train, y_train)
    y_pred = clf.predict(X_test)
    print("\n", clf.best_score_, clf.best_params_)
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred))
    outer_results.append({'f1': f1_score(y_test, y_pred), 'sensitivity': recall_score(y_test, y_pred),
                          'specificity': recall_score(y_test, y_pred, pos_label=0), 'PPV': precision_score(y_test, y_pred),
                          'NPV': precision_score(y_test, y_pred, pos_label=0)})
print("\nMean scores:")
for score in ['f1', 'sensitivity', 'specificity', 'PPV', 'NPV']:
    print(" ", score, ":", np.array([dict[score] for dict in outer_results]).mean())

> Process 2 has produced the most helpful set of results, in particular the highest positive and negative predictive values (class 1 and 0 precision respectively), so this process is repeated on the full dataset using 3-fold cross validation. The resulting model is then saved as two files.

np.random.seed(12)
batch_size = [30, 35, 40, 45]
epochs = [6, 7, 8, 9]
neurons1 = [25, 30, 35, 40, 45, 50]
neurons2 = [30, 40, 50, 60]
dropout1 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
dropout2 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
constraint = [0, 1, 2, 3, 4, 5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
clf = RandomizedSearchCV(estimator=pipeline, param_distributions=param_grid, cv=3, n_jobs=3, verbose=1, n_iter=1000, random_state=1)
clf.fit(X, y)

print(clf.best_score_, clf.best_params_)
fitted_dl_model_ncv = clf.best_estimator_

fitted_dl_model_ncv.named_steps['nn'].model.save('eeg_dl_model_ncv.h5')
fitted_dl_model_ncv.named_steps['nn'].model = None
joblib.dump(fitted_dl_model_ncv, 'eeg_dl_model_ncv.pkl')

> **WITHOUT ABSOLUTE POWERS**
>
> A similar set of models is then produced using a version of the dataset without absolute powers.

y = df.status
X = df.drop(['subject', 'Fp1a delta', 'Fp1a theta', 'Fp1a alpha', 'Fp1a beta',
       'Fp1a gamma', 'Fp2a delta', 'Fp2a theta', 'Fp2a alpha',
       'Fp2a beta', 'Fp2a gamma', 'F3a delta', 'F3a theta', 'F3a alpha',
       'F3a beta', 'F3a gamma', 'F4a delta', 'F4a theta', 'F4a alpha',
       'F4a beta', 'F4a gamma', 'F7a delta', 'F7a theta', 'F7a alpha',
       'F7a beta', 'F7a gamma', 'F8a delta', 'F8a theta', 'F8a alpha',
       'F8a beta', 'F8a gamma', 'C3a delta', 'C3a theta', 'C3a alpha',
       'C3a beta', 'C3a gamma', 'C4a delta', 'C4a theta', 'C4a alpha',
       'C4a beta', 'C4a gamma', 'P3a delta', 'P3a theta', 'P3a alpha',
       'P3a beta', 'P3a gamma', 'P4a delta', 'P4a theta', 'P4a alpha',
       'P4a beta', 'P4a gamma', 'O1a delta', 'O1a theta', 'O1a alpha',
       'O1a beta', 'O1a gamma', 'O2a delta', 'O2a theta', 'O2a alpha',
       'O2a beta', 'O2a gamma', 'status'], axis=1)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=1234, stratify=df.status)

def create_model(neurons1=100, neurons2=30, dropout1=0, dropout2=0, constraint=0):
    model = Sequential()
    model.add(Dense(neurons1, input_dim=90, activation='relu', kernel_constraint=max_norm(constraint)))
    model.add(Dropout(dropout1))
    model.add(Dense(neurons2, activation='relu', kernel_constraint=max_norm(constraint)))
    model.add(Dropout(dropout2))
    model.add(Dense(1, activation='sigmoid'))
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
    return model

pipeline = Pipeline([('scaler', StandardScaler()), ('nn', KerasClassifier(build_fn=create_model, verbose=0))])

# MODEL 1
np.random.seed(12)
batch_size = [30, 35, 40, 45, 50]
epochs = [6, 7, 8, 9, 10, 11, 12]
neurons1 = [25, 37, 50, 75, 100]
neurons2 = [30, 40, 50, 60]
dropout1 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
dropout2 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
constraint = [0, 1, 2, 3, 4, 5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
clf = RandomizedSearchCV(estimator=pipeline, param_distributions=param_grid, cv=3, n_jobs=3, verbose=1, n_iter=1000)
clf.fit(X_train, y_train)

y_pred = (clf.predict(X_test) > 0.5).astype(int)
print(clf.best_score_, clf.best_params_)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

fitted_dl_model_no_ap = clf.best_estimator_

# MODEL 2
np.random.seed(12)
batch_size = [25, 30, 35]
epochs = [8, 9, 10]
neurons1 = [31, 37, 43]
neurons2 = [45, 50, 55]
dropout1 = [0.4, 0.5]
dropout2 = [0, 0.1, 0.2]
constraint = [4, 5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
clf = GridSearchCV(estimator=pipeline, param_grid=param_grid, cv=3, n_jobs=3, verbose=1)
clf.fit(X_train, y_train)

y_pred = (clf.predict(X_test) > 0.5).astype(int)
print(clf.best_score_, clf.best_params_)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

> Model 1 has produced the more clinically useful set of results so this one is saved.

fitted_dl_model_no_ap.named_steps['nn'].model.save('eeg_dl_model_no_ap.h5')
fitted_dl_model_no_ap.named_steps['nn'].model = None
joblib.dump(fitted_dl_model_no_ap, 'eeg_dl_model_no_ap.pkl')

> **Nested cross validation model**

# PROCESS 1
np.random.seed(12)
batch_size = [30, 35, 40, 45, 50]
epochs = [6, 7, 8, 9, 10, 11, 12]
neurons1 = [25, 37, 50, 75, 100]
neurons2 = [30, 40, 50, 60]
dropout1 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
dropout2 = [0, 0.1, 0.2, 0.3, 0.4, 0.5]
constraint = [0, 1, 2, 3, 4, 5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
outer_results = list()
for train_ix, test_ix in cv_outer.split(X, y):
    X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
    y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
    cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
    clf = RandomizedSearchCV(estimator=pipeline, param_distributions=param_grid, cv=3, n_jobs=5, verbose=1, n_iter=1000, random_state=1)
    clf.fit(X_train, y_train)
    y_pred = (clf.predict(X_test) > 0.5).astype(int)
    print("\n", clf.best_score_, clf.best_params_)
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred))
    outer_results.append({'f1': f1_score(y_test, y_pred), 'sensitivity': recall_score(y_test, y_pred),
                          'specificity': recall_score(y_test, y_pred, pos_label=0), 'PPV': precision_score(y_test, y_pred),
                          'NPV': precision_score(y_test, y_pred, pos_label=0)})
print("\nMean scores:")
for score in ['f1', 'sensitivity', 'specificity', 'PPV', 'NPV']:
    print(" ", score, ":", np.array([dict[score] for dict in outer_results]).mean())

# PROCESS 2
batch_size = [30, 37, 45]
epochs = [6, 7]
neurons1 = [20, 25]
neurons2 = [20, 30, 40]
dropout1 = [0, 0.3]
dropout2 = [0.2, 0.5]
constraint = [2, 3, 5]
np.random.seed(12)
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
outer_results = list()
for train_ix, test_ix in cv_outer.split(X, y):
    X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
    y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
    cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
    clf = GridSearchCV(estimator=pipeline, param_grid=param_grid, cv=3, n_jobs=3, verbose=1)
    clf.fit(X_train, y_train)
    y_pred = (clf.predict(X_test) > 0.5).astype(int)
    print("\n", clf.best_score_, clf.best_params_)
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred))
    outer_results.append({'f1': f1_score(y_test, y_pred), 'sensitivity': recall_score(y_test, y_pred),
                          'specificity': recall_score(y_test, y_pred, pos_label=0), 'PPV': precision_score(y_test, y_pred),
                          'NPV': precision_score(y_test, y_pred, pos_label=0)})
print("\nMean scores:")
for score in ['f1', 'sensitivity', 'specificity', 'PPV', 'NPV']:
    print(" ", score, ":", np.array([dict[score] for dict in outer_results]).mean())

# PROCESS 3
np.random.seed(12)
batch_size = [45]
epochs = [6, 7, 8, 9, 10, 11, 12]
neurons1 = [20, 25, 30, 35, 40]
neurons2 = [20, 30, 40]
dropout1 = [0.3]
dropout2 = [0.5]
constraint = [5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
outer_results = list()
for train_ix, test_ix in cv_outer.split(X, y):
    X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
    y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
    cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
    clf = GridSearchCV(estimator=pipeline, param_grid=param_grid, cv=3, n_jobs=3, verbose=1)
    clf.fit(X_train, y_train)
    y_pred = (clf.predict(X_test) > 0.5).astype(int)
    print("\n", clf.best_score_, clf.best_params_)
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred))
    outer_results.append({'f1': f1_score(y_test, y_pred), 'sensitivity': recall_score(y_test, y_pred),
                          'specificity': recall_score(y_test, y_pred, pos_label=0), 'PPV': precision_score(y_test, y_pred),
                          'NPV': precision_score(y_test, y_pred, pos_label=0)})
print("\nMean scores:")
for score in ['f1', 'sensitivity', 'specificity', 'PPV', 'NPV']:
    print(" ", score, ":", np.array([dict[score] for dict in outer_results]).mean())

# PROCESS 4
np.random.seed(12)
batch_size = [30, 35, 40, 45]
epochs = [6, 7, 8, 9]
neurons1 = [20, 25, 30, 35]
neurons2 = [20, 30, 40, 50]
dropout1 = [0, 0.3]
dropout2 = [0.2, 0.5]
constraint = [2, 3, 5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
cv_outer = StratifiedKFold(n_splits=4, shuffle=True, random_state=1)
outer_results = list()
for train_ix, test_ix in cv_outer.split(X, y):
    X_train, X_test = X.iloc[train_ix, :], X.iloc[test_ix, :]
    y_train, y_test = y.iloc[train_ix], y.iloc[test_ix]
    cv_inner = StratifiedKFold(n_splits=3, shuffle=True, random_state=1)
    clf = GridSearchCV(estimator=pipeline, param_grid=param_grid, cv=3, n_jobs=3, verbose=1)
    clf.fit(X_train, y_train)
    y_pred = (clf.predict(X_test) > 0.5).astype(int)
    print("\n", clf.best_score_, clf.best_params_)
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred))
    outer_results.append({'f1': f1_score(y_test, y_pred), 'sensitivity': recall_score(y_test, y_pred),
                          'specificity': recall_score(y_test, y_pred, pos_label=0), 'PPV': precision_score(y_test, y_pred),
                          'NPV': precision_score(y_test, y_pred, pos_label=0)})
print("\nMean scores:")
for score in ['f1', 'sensitivity', 'specificity', 'PPV', 'NPV']:
    print(" ", score, ":", np.array([dict[score] for dict in outer_results]).mean())

> None of these results are suggesting a very strong predictor, but process 2 is the best of these so this is applied to the full dataset using 3-fold cross validation and the resulting model is saved.

np.random.seed(12)
batch_size = [30, 37, 45]
epochs = [6, 7]
neurons1 = [20, 25]
neurons2 = [20, 30, 40]
dropout1 = [0, 0.3]
dropout2 = [0.2, 0.5]
constraint = [2, 3, 5]
param_grid = dict(nn__batch_size=batch_size, nn__epochs=epochs, nn__neurons1=neurons1, nn__neurons2=neurons2,
                 nn__dropout1=dropout1, nn__dropout2=dropout2, nn__constraint=constraint)
clf = RandomizedSearchCV(estimator=pipeline, param_distributions=param_grid, cv=3, n_jobs=3, verbose=1, n_iter=1000, random_state=1)
clf.fit(X, y)

print(clf.best_score_, clf.best_params_)
fitted_dl_model_no_ap_ncv = clf.best_estimator_

fitted_dl_model_no_ap_ncv.named_steps['nn'].model.save('eeg_dl_model_no_ap_ncv.h5')
fitted_dl_model_no_ap_ncv.named_steps['nn'].model = None
joblib.dump(fitted_dl_model_no_ap_ncv, 'eeg_dl_model_no_ap_ncv.pkl')
