==> 5 Model testing.ipynb <==
> In this notebook the models are tested on the MPIL dataset. Required libraries and files from the dataset are imported. A separate DataFrame has been compiled for each of the two testing conditions in the MPIL study &ndash; `dfeo` for eyes open, and `dfec` for eyes closed.

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import pickle
import joblib
from sklearn.metrics import confusion_matrix, f1_score, classification_report
from keras.models import Sequential, load_model
from keras.layers import Dense, Dropout
from keras.constraints import max_norm
from keras.wrappers.scikit_learn import KerasClassifier

dfeo = pd.read_csv('EEG_MPIL_dataset_powers_EO.csv')
dfec = pd.read_csv('EEG_MPIL_dataset_powers_EC.csv')
dfsd = pd.read_csv('MPIL_subject_data.csv')

dfeo

dfec

dfsd

> The AUDIT screen is a questionnaire on alcohol use that generates a score ranging from 0 to 40. A score of 8 or above indicates hazardous or risky drinking, so this has been used as a cut-off for dividing the participant group into hazardous drinkers and others. In the cell below a label column is added to the `dfeo` and `dfec` DataFrames, showing "1" for hazardous drinkers and "0" for others. Any subjects not appearing in all 3 DataFrames are dropped from the analysis.

dfeo['label'] = [0] * len(dfeo.index)
dfec['label'] = [0] * len(dfec.index)
for sub_id in dfeo['Unnamed: 0']:
    if not sub_id in dfec['Unnamed: 0'].values:
        dfeo = dfeo.drop(dfeo.index[dfeo['Unnamed: 0'] == sub_id].item())
    else:
        if not "sub-0" + str(int(sub_id)) in dfsd['ID'].values:
            dfeo = dfeo.drop(dfeo.index[dfeo['Unnamed: 0'] == sub_id].item())
            dfec = dfec.drop(dfeo.index[dfec['Unnamed: 0'] == sub_id].item())
        else:
            idxsd = dfsd.index[dfsd['ID'] == "sub-0" + str(int(sub_id))].item()
            idxeo = dfeo.index[dfeo['Unnamed: 0'] == sub_id].item()
            idxec = dfec.index[dfec['Unnamed: 0'] == sub_id].item()
            label = int(dfsd.loc[idxsd]['AUDIT'] >= 8.0)
            dfeo.loc[idxeo, 'label'] = label
            dfec.loc[idxec, 'label'] = label
for sub_id in dfec['Unnamed: 0']:
    if not sub_id in dfeo['Unnamed: 0'].values:
        dfec = dfec.drop(dfec.index[dfec['Unnamed: 0'] == sub_id].item())

dfeo

dfec

X_eo = dfeo.drop(['Unnamed: 0', 'label'], axis=1)
X_ec = dfec.drop(['Unnamed: 0', 'label'], axis=1)
y_eo = dfeo['label']
y_ec = dfeo['label']

> Saved models are retrieved below. The `create_model` function needs to be in place for the deep learning models to load correctly.

fitted_model = pickle.load(open('eeg_model.pkl', 'rb'))
fitted_model_ncv = pickle.load(open('eeg_model_ncv.pkl', 'rb'))
best_thr_ncv = pickle.load(open('eeg_best_thr_ncv.pkl', 'rb'))

def create_model(neurons1=100, neurons2=30, dropout1=0, dropout2=0, constraint=0):
    model = Sequential()
    model.add(Dense(neurons1, input_dim=150, activation='relu', kernel_constraint=max_norm(constraint)))
    model.add(Dropout(dropout1))
    model.add(Dense(neurons2, activation='relu', kernel_constraint=max_norm(constraint)))
    model.add(Dropout(dropout2))
    model.add(Dense(1, activation='sigmoid'))
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
    return model

fitted_dl_model = joblib.load('eeg_dl_model.pkl')
fitted_dl_model.named_steps['nn'].model = load_model('eeg_dl_model.h5')

fitted_dl_model_ncv = joblib.load('eeg_dl_model_ncv.pkl')
fitted_dl_model_ncv.named_steps['nn'].model = load_model('eeg_dl_model_ncv.h5')

> The various models are then tested on the two subsets of MPIL data and metrics are displayed. 
>
> The confusion matrix in each case is formatted as follows:  
> 
> &nbsp; &nbsp; \[[ true negatives | false positives ]  
> &nbsp; &nbsp; &nbsp;[ false negatives | true positives ]]  
> 
> The key variables of interest for clinical purposes are:
> - sensitivity (= class 1 recall) &ndash; the probablility of correctly identifying a positive (alcoholic) case
> - specificity (= class 0 recall) &ndash; the probablility of correctly identifying a negative (non-alcoholic) case
>
> Note that if sensitivity and specificity add up to 1, the predictor is making absolutely no distinction between positive and negative cases.

# MODEL 1 - L2 model, eyes open
pred = fitted_model.predict(X_eo)
print(confusion_matrix(y_eo, pred))
print(classification_report(y_eo, pred))

# MODEL 2 - Nested cross validation (NCV) derived L2 model, eyes open
pred = fitted_model_ncv.predict(X_eo)
print(confusion_matrix(y_eo, pred))
print(classification_report(y_eo, pred))

# MODEL 3 - NCV-derived L2 model with optimised threshold, eyes open
try:
    pred = fitted_model_ncv.predict_proba(X_eo)
    pred = pred[:, 1]
except:
    pred = fitted_model_ncv.decision_function(X_eo)
y_pred = [int(p > best_thr_ncv) for p in pred]
print(confusion_matrix(y_eo, y_pred))
print(classification_report(y_eo, y_pred))

# MODEL 4 - deep learning model, eyes open
pred = (fitted_dl_model.predict(X_eo) > 0.5).astype(int)
print(confusion_matrix(y_eo, pred))
print(classification_report(y_eo, pred))

# MODEL 5 - NCV-derived deep learning model, eyes open
pred = (fitted_dl_model_ncv.predict(X_eo) > 0.5).astype(int)
print(confusion_matrix(y_eo, pred))
print(classification_report(y_eo, pred))

# MODEL 6 - L2 model, eyes closed
pred = fitted_model.predict(X_ec)
print(confusion_matrix(y_ec, pred))
print(classification_report(y_ec, pred))

# MODEL 7 - NCV-derived L2 model, eyes closed
pred = fitted_model_ncv.predict(X_ec)
print(confusion_matrix(y_ec, pred))
print(classification_report(y_ec, pred))

# MODEL 8 - NCV-derived L2 model with optimised threshold, eyes closed
try:
    pred = fitted_model_ncv.predict_proba(X_ec)
    pred = pred[:, 1]
except:
    pred = fitted_model_ncv.decision_function(X_ec)
y_pred = [int(p > best_thr_ncv) for p in pred]
print(confusion_matrix(y_ec, y_pred))
print(classification_report(y_ec, y_pred))

# MODEL 9 - deep learning model, eyes closed
pred = (fitted_dl_model.predict(X_ec) > 0.5).astype(int)
print(confusion_matrix(y_ec, pred))
print(classification_report(y_ec, pred))

# MODEL 10 - NCV-derived deep learning model, eyes closed
pred = (fitted_dl_model_ncv.predict(X_ec) > 0.5).astype(int)
print(confusion_matrix(y_ec, pred))
print(classification_report(y_ec, pred))

> **WITHOUT ABSOLUTE POWERS**
> 
> The MPIL datasets are modified by removing absolute powers, and these are then used for testing models trained on UCI data without absolute powers.

X_eo_no_ap = dfeo.drop(['Unnamed: 0', 'Fp1a delta', 'Fp1a theta', 'Fp1a alpha', 'Fp1a beta',
       'Fp1a gamma', 'Fp2a delta', 'Fp2a theta', 'Fp2a alpha',
       'Fp2a beta', 'Fp2a gamma', 'F3a delta', 'F3a theta', 'F3a alpha',
       'F3a beta', 'F3a gamma', 'F4a delta', 'F4a theta', 'F4a alpha',
       'F4a beta', 'F4a gamma', 'F7a delta', 'F7a theta', 'F7a alpha',
       'F7a beta', 'F7a gamma', 'F8a delta', 'F8a theta', 'F8a alpha',
       'F8a beta', 'F8a gamma', 'C3a delta', 'C3a theta', 'C3a alpha',
       'C3a beta', 'C3a gamma', 'C4a delta', 'C4a theta', 'C4a alpha',
       'C4a beta', 'C4a gamma', 'P3a delta', 'P3a theta', 'P3a alpha',
       'P3a beta', 'P3a gamma', 'P4a delta', 'P4a theta', 'P4a alpha',
       'P4a beta', 'P4a gamma', 'O1a delta', 'O1a theta', 'O1a alpha',
       'O1a beta', 'O1a gamma', 'O2a delta', 'O2a theta', 'O2a alpha',
       'O2a beta', 'O2a gamma', 'label'], axis=1)

X_ec_no_ap = dfec.drop(['Unnamed: 0', 'Fp1a delta', 'Fp1a theta', 'Fp1a alpha', 'Fp1a beta',
       'Fp1a gamma', 'Fp2a delta', 'Fp2a theta', 'Fp2a alpha',
       'Fp2a beta', 'Fp2a gamma', 'F3a delta', 'F3a theta', 'F3a alpha',
       'F3a beta', 'F3a gamma', 'F4a delta', 'F4a theta', 'F4a alpha',
       'F4a beta', 'F4a gamma', 'F7a delta', 'F7a theta', 'F7a alpha',
       'F7a beta', 'F7a gamma', 'F8a delta', 'F8a theta', 'F8a alpha',
       'F8a beta', 'F8a gamma', 'C3a delta', 'C3a theta', 'C3a alpha',
       'C3a beta', 'C3a gamma', 'C4a delta', 'C4a theta', 'C4a alpha',
       'C4a beta', 'C4a gamma', 'P3a delta', 'P3a theta', 'P3a alpha',
       'P3a beta', 'P3a gamma', 'P4a delta', 'P4a theta', 'P4a alpha',
       'P4a beta', 'P4a gamma', 'O1a delta', 'O1a theta', 'O1a alpha',
       'O1a beta', 'O1a gamma', 'O2a delta', 'O2a theta', 'O2a alpha',
       'O2a beta', 'O2a gamma', 'label'], axis=1)

fitted_model_no_ap = pickle.load(open('eeg_model_no_ap.pkl', 'rb'))
fitted_model_no_ap_ncv = pickle.load(open('eeg_model_no_ap_ncv.pkl', 'rb'))
best_thr_no_ap_ncv = pickle.load(open('eeg_best_thr_no_ap_ncv.pkl', 'rb'))

def create_model(neurons1=100, neurons2=30, dropout1=0, dropout2=0, constraint=0):
    model = Sequential()
    model.add(Dense(neurons1, input_dim=90, activation='relu', kernel_constraint=max_norm(constraint)))
    model.add(Dropout(dropout1))
    model.add(Dense(neurons2, activation='relu', kernel_constraint=max_norm(constraint)))
    model.add(Dropout(dropout2))
    model.add(Dense(1, activation='sigmoid'))
    model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
    return model

fitted_dl_model_no_ap = joblib.load('eeg_dl_model_no_ap.pkl')
fitted_dl_model_no_ap.named_steps['nn'].model = load_model('eeg_dl_model_no_ap.h5')

fitted_dl_model_no_ap_ncv = joblib.load('eeg_dl_model_no_ap_ncv.pkl')
fitted_dl_model_no_ap_ncv.named_steps['nn'].model = load_model('eeg_dl_model_no_ap_ncv.h5')

# MODEL 11 - linear SVM model, eyes open
pred = fitted_model_no_ap.predict(X_eo_no_ap)
print(confusion_matrix(y_eo, pred))
print(classification_report(y_eo, pred))

# MODEL 12 - NCV-derived L1 model, eyes open
pred = fitted_model_no_ap_ncv.predict(X_eo_no_ap)
print(confusion_matrix(y_eo, pred))
print(classification_report(y_eo, pred))

# MODEL 13 - NCV-derived L1 model with optimised threshold, eyes open
try:
    pred = fitted_model_no_ap_ncv.predict_proba(X_eo_no_ap)
    pred = pred[:, 1]
except:
    pred = fitted_model_no_ap_ncv.decision_function(X_eo_no_ap)
y_pred = [int(p > best_thr_ncv) for p in pred]
print(confusion_matrix(y_eo, y_pred))
print(classification_report(y_eo, y_pred))

# MODEL 14 - deep learning model, eyes open
pred = (fitted_dl_model_no_ap.predict(X_eo_no_ap) > 0.5).astype(int)
print(confusion_matrix(y_eo, pred))
print(classification_report(y_eo, pred))

# MODEL 15 - NCV-derived deep learning model, eyes open
# This has slight utility as a rule-out test
pred = (fitted_dl_model_no_ap_ncv.predict(X_eo_no_ap) > 0.5).astype(int)
print(confusion_matrix(y_eo, pred))
print(classification_report(y_eo, pred))

# MODEL 16 - linear SVM model, eyes closed
pred = fitted_model_no_ap.predict(X_ec_no_ap)
print(confusion_matrix(y_ec, pred))
print(classification_report(y_ec, pred))

# MODEL 17 - NCV-derived L1 model, eyes closed
pred = fitted_model_no_ap_ncv.predict(X_ec_no_ap)
print(confusion_matrix(y_ec, pred))
print(classification_report(y_ec, pred))

# MODEL 18 - NCV-derived L1 model with optimised threshold, eyes closed
try:
    pred = fitted_model_no_ap_ncv.predict_proba(X_ec_no_ap)
    pred = pred[:, 1]
except:
    pred = fitted_model_no_ap_ncv.decision_function(X_ec_no_ap)
y_pred = [int(p > best_thr_ncv) for p in pred]
print(confusion_matrix(y_ec, y_pred))
print(classification_report(y_ec, y_pred))

# MODEL 19 - deep learning model, eyes closed
pred = (fitted_dl_model_no_ap.predict(X_ec_no_ap) > 0.5).astype(int)
print(confusion_matrix(y_ec, pred))
print(classification_report(y_ec, pred))

# MODEL 20 - NCV-derived deep learning model, eyes closed
pred = (fitted_dl_model_no_ap_ncv.predict(X_ec_no_ap) > 0.5).astype(int)
print(confusion_matrix(y_ec, pred))
print(classification_report(y_ec, pred))

> Above we have 10 different models each tested on 2 datasets in order to give 20 tests in total. In most of the cases sensitivity and specificity add up to less than 1, meaning that the usefulness of the predictor is worse than a coin toss. A positive result for an individual subject from any of these predictors is of no real value, as the highest positive predictive value (class 1 precision, indicating the probability that a positive result is correct) is 0.21. Note that in the test sets we are using, the proportion of alcoholic subjects is 0.17, so the negative predictive value (class 0 precision) of a completely random coin toss as a test for alcoholism in this subject group (i.e. the chance of a *negative* result being correct) would be 0.83. The only models with a negative predictive higher than this are some of the deep learning models.
>
> Model 15 would have slight utility as a rule-out test. Class 0 precision is 0.91, meaning that 91% of negative test results would be correct, whereas with a class 1 precision of only 0.19, a positive test result would be of no real use. Recall values for the two classes show that 89% of hazardous drinkers are identified by the test as such, but then so are 79% of people who are not hazardous drinkers. The usefulness of model 15 as a predictor should be interpreted very cautiously, as one seemingly helpful result out of 20 tests could easily occur by chance.
>
> Overall, the models trained on the UCI dataset have not proven helpful for predicting hazardous drinking when tested on the MPIL dataset. It may still be possible to train some helpful models for this purpose &ndash; we don't know for sure either way &ndash; but this might require a different approach to calculating band powers, an approach other than band power, more rigorous artefact removal, or a test set with testing conditions and a case definition more closely matching those of the training set.
